requirements.txt
stravalib
stravaweblib
garminconnect
thefuzz
pillow
python-dotenv